// src/models/LoginStudent.ts
export default interface LoginStudent {
    id: number;
    ra: string;
    name: string;
    email: string;
    rg: string;
    cpf: string;
    course: string;
    period: string;
    status: string;
    admission: string;
    birthDate: string;
    dueDate: string;
    photo: string;
    qrcode: string;
}